requirements.txt
discord.py==1.7.3
python-dotenv==0.20.0
selenium==4.1.5
validators==0.19.0
